Break app.js into separate sections
 - app.js is getting hard to understand, put each section into it's own
   file
